break up the simulator primitives into their own files in the simulator submodule
